implemented "back to the station" button

+++ view/main_menu_view.py
@@ -6,6 +6,7 @@ from pyshaders import from_files_names
 
 from view import *
 from ui.button.create_station_button import CreateStationButton
+from ui.button.back_to_the_station_button import BackToTheStationButton
 from ui.button.open_license_button import OpenLicenseButton
 from i18n import I18N_RESOURCES
 
@@ -19,13 +20,16 @@ class MainMenuView(View):
         """
         Button click handlers:
             on_create_station                   on_click handler for create station button
+            on_back_to_the_station              on_click handler for back to the station button
             on_open_license                     on_click handler for open license button
 
         Properties:
             create_station_button               CreateStationButton object
+            back_to_the_station_button          BackToTheStationButton
             open_license_button                 OpenLicenseButton object
             buttons                             list of all buttons
             create_station_button_label         label from "Create new station" title
+            back_to_the_station_label           label from "Back to the station" title
 
         """
         def on_create_station(button):
@@ -40,6 +44,19 @@ class MainMenuView(View):
             self.controller.parent_controller.game_to_main_menu_transition_animation.on_deactivate()
             self.controller.parent_controller.main_menu_to_onboarding_transition_animation.on_activate()
 
+        def on_back_to_the_station(button):
+            """
+            Performs transition from main menu screen to game screen.
+
+            :param button:                      button that was clicked
+            """
+            button.on_deactivate()
+            self.controller.on_deactivate_view()
+            self.controller.parent_controller.license_to_main_menu_transition_animation.on_deactivate()
+            self.controller.parent_controller.game_to_main_menu_transition_animation.on_deactivate()
+            self.controller.parent_controller.main_menu_to_game_transition_animation.on_activate()
+            self.controller.parent_controller.on_resume_game()
+
         def on_open_license(button):
             """
             Performs transition from main menu screen to license screen.
@@ -51,9 +68,11 @@ class MainMenuView(View):
 
         super().__init__(logger=getLogger('root.app.main_menu.view'))
         self.create_station_button = CreateStationButton(on_click_action=on_create_station)
+        self.back_to_the_station_button = BackToTheStationButton(on_click_action=on_back_to_the_station)
         self.open_license_button = OpenLicenseButton(on_click_action=on_open_license)
-        self.buttons = [self.create_station_button, self.open_license_button]
+        self.buttons = [self.create_station_button, self.back_to_the_station_button, self.open_license_button]
         self.create_station_button_label = None
+        self.back_to_the_station_label = None
         self.shader = from_files_names('shaders/shader.vert', 'shaders/main_menu_view/shader.frag')
         self.shader_sprite = None
         self.on_init_graphics()
@@ -82,10 +101,20 @@ class MainMenuView(View):
         if self.opacity <= 0:
             self.shader_sprite.delete()
             self.shader_sprite = None
-            self.create_station_button_label.delete()
-            self.create_station_button_label = None
+            if self.create_station_button_label is not None:
+                self.create_station_button_label.delete()
+                self.create_station_button_label = None
+
+            if self.back_to_the_station_label is not None:
+                self.back_to_the_station_label.delete()
+                self.back_to_the_station_label = None
+
         else:
-            self.create_station_button_label.color = (*WHITE_RGB, self.opacity)
+            if self.create_station_button_label is not None:
+                self.create_station_button_label.color = (*WHITE_RGB, self.opacity)
+
+            if self.back_to_the_station_label is not None:
+                self.back_to_the_station_label.color = (*WHITE_RGB, self.opacity)
 
     @view_is_not_active
     def on_activate(self):
@@ -98,14 +127,28 @@ class MainMenuView(View):
                 = self.batches['main_frame'].add(4, GL_QUADS, self.groups['main_frame'],
                                                  ('v2f/static', (-1.0, -1.0, -1.0, 1.0, 1.0, 1.0, 1.0, -1.0)))
 
-        if self.create_station_button_label is None:
-            self.create_station_button_label \
-                = Label(I18N_RESOURCES['create_station_label_string'][self.current_locale],
-                        font_name='Perfo', bold=True, font_size=3 * self.bottom_bar_height // 8,
-                        color=(*WHITE_RGB, self.opacity), x=self.screen_resolution[0] // 2,
-                        y=self.screen_resolution[1] // 2 + int(72 / 1280 * self.screen_resolution[0]) // 4,
-                        anchor_x='center', anchor_y='center',
-                        batch=self.batches['ui_batch'], group=self.groups['button_text'])
+        self.user_db_cursor.execute('SELECT onboarding_required FROM game_progress')
+        onboarding_required = bool(self.user_db_cursor.fetchone()[0])
+        if onboarding_required:
+            self.create_station_button.on_activate()
+            if self.create_station_button_label is None:
+                self.create_station_button_label \
+                    = Label(I18N_RESOURCES['create_station_label_string'][self.current_locale],
+                            font_name='Perfo', bold=True, font_size=3 * self.bottom_bar_height // 8,
+                            color=(*WHITE_RGB, self.opacity), x=self.screen_resolution[0] // 2,
+                            y=self.screen_resolution[1] // 2 + int(72 / 1280 * self.screen_resolution[0]) // 4,
+                            anchor_x='center', anchor_y='center',
+                            batch=self.batches['ui_batch'], group=self.groups['button_text'])
+        else:
+            self.back_to_the_station_button.on_activate()
+            if self.back_to_the_station_label is None:
+                self.back_to_the_station_label \
+                    = Label(I18N_RESOURCES['back_to_the_station_label_string'][self.current_locale],
+                            font_name='Perfo', bold=True, font_size=3 * self.bottom_bar_height // 8,
+                            color=(*WHITE_RGB, self.opacity), x=self.screen_resolution[0] // 2,
+                            y=self.screen_resolution[1] // 2 + int(72 / 1280 * self.screen_resolution[0]) // 4,
+                            anchor_x='center', anchor_y='center',
+                            batch=self.batches['ui_batch'], group=self.groups['button_text'])
 
         for b in self.buttons:
             if b.to_activate_on_controller_init:
@@ -130,14 +173,23 @@ class MainMenuView(View):
         self.on_recalculate_ui_properties(screen_resolution)
         medium_line = self.screen_resolution[1] // 2 + int(72 / 1280 * self.screen_resolution[0]) // 4
         if self.is_activated:
-            self.create_station_button_label.x = self.screen_resolution[0] // 2
-            self.create_station_button_label.y = medium_line
-            self.create_station_button_label.font_size = 3 * self.bottom_bar_height // 8
+            if self.create_station_button_label is not None:
+                self.create_station_button_label.x = self.screen_resolution[0] // 2
+                self.create_station_button_label.y = medium_line
+                self.create_station_button_label.font_size = 3 * self.bottom_bar_height // 8
+
+            if self.back_to_the_station_label is not None:
+                self.back_to_the_station_label.x = self.screen_resolution[0] // 2
+                self.back_to_the_station_label.y = medium_line
+                self.back_to_the_station_label.font_size = 3 * self.bottom_bar_height // 8
 
         self.open_license_button.on_size_changed((self.bottom_bar_height, self.bottom_bar_height))
         self.create_station_button.on_size_changed((self.bottom_bar_height * 7, self.bottom_bar_height))
         self.create_station_button.x_margin = self.screen_resolution[0] // 2 - self.bottom_bar_height * 7 // 2
         self.create_station_button.y_margin = medium_line - self.bottom_bar_height // 2
+        self.back_to_the_station_button.on_size_changed((self.bottom_bar_height * 7, self.bottom_bar_height))
+        self.back_to_the_station_button.x_margin = self.screen_resolution[0] // 2 - self.bottom_bar_height * 7 // 2
+        self.back_to_the_station_button.y_margin = medium_line - self.bottom_bar_height // 2
         for b in self.buttons:
             b.on_position_changed((b.x_margin, b.y_margin))
 
@@ -149,7 +201,13 @@ class MainMenuView(View):
         """
         self.current_locale = new_locale
         if self.is_activated:
-            self.create_station_button_label.text = I18N_RESOURCES['create_station_label_string'][self.current_locale]
+            if self.create_station_button_label is not None:
+                self.create_station_button_label.text \
+                    = I18N_RESOURCES['create_station_label_string'][self.current_locale]
+
+            if self.back_to_the_station_label is not None:
+                self.back_to_the_station_label.text \
+                    = I18N_RESOURCES['back_to_the_station_label_string'][self.current_locale]
 
     @shader_sprite_exists
     def on_apply_shaders_and_draw_vertices(self):
